MockDeviceStorage should only available in iOS
